Move licenses plugin to apputils-extension
