+++ packages/apputils/style/base.css
@@ -6,6 +6,7 @@
 @import './commandpalette.css';
 @import './dialog.css';
 @import './inputdialog.css';
+@import './licenses.css';
 @import './mainareawidget.css';
 @import './materialcolors.css';
 @import './toolbar.css';
